// src/modules/app/components/routes/types.ts
export enum Paths {
    home = "/",
    dashboard = "dashboard",
    pool = "pool",
    pooldetails = "pool/pooldetails",
    lenders = "lenders",
    borrow = "borrow",
    stake = "stake",
    claimRewards = "claim-rewards",
    liquidity = "liquidity",
    mint="/mint",
    claimldf="/claimldf",
    deposit="pool/pooldetails/deposit",
    borrowing="pool/pooldetails/borrow"
  }